pyproject.toml
[tool.poetry]
name = "oi-ai-mleng-takehome"
version = "0.1.0"
description = ""
authors = ["Bruno Aguiar <bruno.f.o.aguiar@gmail.com>"]
readme = "README.md"
packages = [{include = "oi"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.1"
torchvision = "^0.18.1"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
aiohttp = "^3.9.5"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
invoke = "^2.2.0"
httpx = "^0.27.0"
pylint = "^3.2.5"
isort = "^5.13.2"
black = "^24.4.2"

[tool.black]
line-length = 79
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "C0114", # Missing module docstring
    "C0116", # Missing function or method docstring
    "R0903", # Too few public methods
    "W0511", # TODO found
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
